Don't run the main loop from __init__
Instead of calling the main loop directly from the class' initializer,
we make a small wrapper function to create the Game object and call
the main loop explicitly. This way it's possible to create a Game object
without running the main loop, and do some testing or monkey patching on
it first.